+++ dispatcher/src/main/java/org/memorizing/resource/StorageResourceImpl.java
@@ -17,7 +17,7 @@ public class StorageResourceImpl implements StorageResource {
     // /storage
 
     @Override
-    public StorageDto getStorageByUserId(Integer userId) {
+    public StorageDto getStorageByUserId(Long userId) {
         return cardWebClientBuilder.getStorageByUserId(userId);
     }
